AP-4575: Namespace copy_case controllers
