+++ app/controllers/providers/copy_case/invitations_controller.rb
@@ -1,0 +1,27 @@
+module Providers
+  module CopyCase
+    class InvitationsController < ProviderBaseController
+      prefix_step_with :copy_case
+
+      def show
+        @form = ::CopyCase::InvitationForm.new(model: legal_aid_application)
+      end
+
+      def update
+        @form = ::CopyCase::InvitationForm.new(form_params)
+
+        render :show unless save_continue_or_draft(@form)
+      end
+
+    private
+
+      def form_params
+        merge_with_model(legal_aid_application) do
+          next {} unless params[:legal_aid_application]
+
+          params.require(:legal_aid_application).permit(:copy_case)
+        end
+      end
+    end
+  end
+end
