+++ app/views/providers/copy_case/confirmations/show.html.erb
