+++ app/views/providers/copy_case/invitations/show.html.erb
