+++ app/views/providers/copy_case/searches/show.html.erb
