+++ config/routes.rb
@@ -160,8 +160,11 @@ Rails.application.routes.draw do
         resources :remove_state_benefits, only: %i[show update]
       end
       get :search, on: :collection
-      resource :copy_case_invitation, only: %i[show update]
-      resource :copy_case_search, only: %i[show update]
+      namespace :copy_case do
+        resource :invitation, only: %i[show update]
+        resource :search, only: %i[show update]
+        resource :confirmation, only: %i[show update]
+      end
       resource :copy_case_confirmation, only: %i[show update]
       resource :delete, controller: :delete, only: %i[show destroy]
       resources :proceedings_types, only: %i[index create]
