+++ spec/requests/providers/copy_case/searches_controller_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe Providers::CopyCaseSearchesController do
+RSpec.describe Providers::CopyCase::SearchesController do
   let(:legal_aid_application) { create(:legal_aid_application) }
   let(:provider) { legal_aid_application.provider }
   let(:login) { login_as provider }
